requirements.txt
streamlit
langchain_google_genai
langchain_community
langchain-openai
langchain
langgraph
youtube-transcript-api
google-api-python-client
bs4
python-pptx
icrawler
openai
langchain_core
langchain-nomic
tiktoken
langchainhub
chromadb
tavily-python
gpt4all
pypdf
PyPDF2
yfinance
